NAS-XXXXXX: Move replace-disk-dialog to storage2
