+++ src/app/pages/storage/volumes/volume-status/volume-status.component.ts
@@ -33,16 +33,13 @@ import { EntityJobComponent } from 'app/modules/entity/entity-job/entity-job.com
 import { EntityTreeTable } from 'app/modules/entity/entity-tree-table/entity-tree-table.model';
 import { EntityUtils } from 'app/modules/entity/utils';
 import { DiskFormComponent } from 'app/pages/storage/disks/disk-form/disk-form.component';
+import { ReplaceDiskDialogComponent, ReplaceDiskDialogData } from 'app/pages/storage2/modules/disks/components/replace-disk-dialog/replace-disk-dialog.component';
 import {
   WebSocketService, AppLoaderService, DialogService,
 } from 'app/services';
 import { IxSlideInService } from 'app/services/ix-slide-in.service';
 import { LayoutService } from 'app/services/layout.service';
 import { ModalService } from 'app/services/modal.service';
-import {
-  ReplaceDiskDialogData,
-  ReplaceDiskDialogComponent,
-} from './components/replace-disk-dialog/replace-disk-dialog.component';
 
 @UntilDestroy()
 @Component({
