+++ src/app/pages/storage2/modules/devices/components/disk-info-card/disk-info-card.component.spec.ts
@@ -10,9 +10,9 @@ import { of } from 'rxjs';
 import { DiskStandby } from 'app/enums/disk-standby.enum';
 import { DiskType } from 'app/enums/disk-type.enum';
 import { Disk } from 'app/interfaces/storage.interface';
-import { DiskFormComponent } from 'app/pages/storage/disks/disk-form/disk-form.component';
-import { ReplaceDiskDialogComponent } from 'app/pages/storage/volumes/volume-status/components/replace-disk-dialog/replace-disk-dialog.component';
 import { DevicesStore } from 'app/pages/storage2/modules/devices/stores/devices-store.service';
+import { DiskFormComponent } from 'app/pages/storage2/modules/disks/components/disk-form/disk-form.component';
+import { ReplaceDiskDialogComponent } from 'app/pages/storage2/modules/disks/components/replace-disk-dialog/replace-disk-dialog.component';
 import { IxSlideInService } from 'app/services/ix-slide-in.service';
 import { DiskInfoCardComponent } from './disk-info-card.component';
