+++ src/app/pages/storage2/modules/disks/components/replace-disk-dialog/replace-disk-dialog.component.scss
